chore(models): Moves re-structured config and model bases to their correct files, and asserts that model_dict and model_loader still works.
