+++ archai/nlp/models/model_dict.py
@@ -4,26 +4,30 @@
 """Availability dictionaries of implemented Transformer-based classes.
 """
 
-from archai.nlp.models.hf_gpt2.config_hf_gpt2 import HfGPT2Config, HfGPT2FlexConfig
+# Huggingface's Open AI GPT-2
+from archai.nlp.models.hf_gpt2.config_hf_gpt2 import (HfGPT2Config, HfGPT2SearchConfig,
+                                                      HfGPT2FlexConfig, HfGPT2FlexSearchConfig)
 from archai.nlp.models.hf_gpt2.model_hf_gpt2 import HfGPT2, HfGPT2Flex
 from archai.nlp.models.hf_gpt2.onnx_hf_gpt2 import HfGPT2OnnxConfig, HfGPT2OnnxModel
 
+# Huggingface's Transformer-XL
 from archai.nlp.models.hf_transfo_xl.config_hf_transfo_xl import HfTransfoXLConfig
 from archai.nlp.models.hf_transfo_xl.model_hf_transfo_xl import HfTransfoXL
 from archai.nlp.models.hf_transfo_xl.onnx_hf_transfo_xl import (HfTransfoXLOnnxConfig,
                                                                 HfTransfoXLOnnxModel)
 
+# NVIDIA's Memory Transformer
 from archai.nlp.models.mem_transformer.config_mem_transformer import MemTransformerLMConfig
 from archai.nlp.models.mem_transformer.model_mem_transformer import MemTransformerLM
 from archai.nlp.models.mem_transformer.onnx_mem_transformer import (MemTransformerLMOnnxConfig,
                                                                     MemTransformerLMOnnxModel)
 
+# Analytical parameters formulae
 from archai.nlp.models.model_utils.analytical_params_formulae import (get_params_hf_gpt2_formula,
                                                                       get_params_hf_gpt2_flex_formula,
                                                                       get_params_hf_transfo_xl_formula,
                                                                       get_params_mem_transformer_formula)
 
-# Available models, their configurations and analytical parameters formulae
 MODELS = {
     'hf_gpt2': HfGPT2,
     'hf_gpt2_flex': HfGPT2Flex,
@@ -38,6 +42,13 @@ MODELS_CONFIGS = {
     'mem_transformer': MemTransformerLMConfig
 }
 
+MODELS_SEARCH_CONFIGS = {
+    'hf_gpt2': HfGPT2SearchConfig,
+    'hf_gpt2_flex': HfGPT2FlexSearchConfig,
+    'hf_transfo_xl': None,
+    'mem_transformer': None
+}
+
 MODELS_PARAMS_FORMULAE = {
     'hf_gpt2': get_params_hf_gpt2_formula,
     'hf_gpt2_flex': get_params_hf_gpt2_flex_formula,
@@ -45,7 +56,6 @@ MODELS_PARAMS_FORMULAE = {
     'mem_transformer': get_params_mem_transformer_formula
 }
 
-# Available ONNX-based models and their configurations
 ONNX_MODELS = {
     'hf_gpt2': HfGPT2OnnxModel,
     'hf_gpt2_flex': HfGPT2OnnxModel,
